refactor: localize startup compat consumers
