+++ docs/architecture/migration-progress.md
@@ -6,15 +6,15 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
 - Branch: `overtrue/arch-runtime-compat-consumer-batch`
-- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097`.
+- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099`.
 - Stacked on: `overtrue/arch-disk-rpc-method-wrappers` pending API-097 merge.
 - PR type for this branch: `pure-move`
 - Runtime behavior changes: none.
-- Rust code changes: move capacity and server ECStore compatibility consumers
-  from the root runtime facade into local capacity/server boundaries.
-- CI/script changes: guard against restoring capacity/server consumer wrappers
+- Rust code changes: move capacity, server, and startup ECStore compatibility
+  consumers from the root runtime facade into local boundaries.
+- CI/script changes: guard against restoring capacity/server/startup consumer wrappers
   in the root runtime compatibility facade.
-- Docs changes: record the API-098 runtime consumer boundary cleanup.
+- Docs changes: record the API-098/API-099 runtime consumer boundary cleanup.
 
 ## Phase 0 Tasks
 
@@ -402,6 +402,22 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
     storage readiness, and distributed lock quorum behavior.
   - Verification: RustFS test-target compile coverage, capacity/server residual
     scan, migration and layer guards, formatting, diff hygiene, Rust risk scan,
+    pre-commit quality gate, and three-expert review.
+- [x] `API-099` Prune root runtime startup compat consumers.
+  - Completed slice: move startup storage bootstrap, bucket metadata migration,
+    notification initialization, global region/port setup, background shutdown,
+    and startup service ECStore aliases into a dedicated startup compatibility
+    boundary.
+  - Acceptance: startup and init modules no longer consume root
+    `storage_compat.rs`; root runtime `storage_compat.rs` no longer owns
+    startup-only ECStore wrapper functions or aliases; migration rules reject
+    restoring those wrappers to the root facade.
+  - Must preserve: endpoint parsing, unsupported filesystem policy,
+    local-disk and lock-client initialization, global config migration,
+    bucket metadata migration, IAM migration, notification registration,
+    default-region fallback, background replication, and shutdown behavior.
+  - Verification: RustFS test-target compile coverage, startup residual scan,
+    migration and layer guards, formatting, diff hygiene, Rust risk scan,
     pre-commit quality gate, and three-expert review.
 - [x] `G-012` Inventory placement and repair invariants.
   - Acceptance:
@@ -3453,6 +3469,19 @@ Passed before push:
   - `./scripts/check_architecture_migration_rules.sh`: passed.
   - `./scripts/check_layer_dependencies.sh`: passed.
   - Direct root capacity/server compatibility consumer residual scan: passed.
+  - Rust risk scan on changed Rust files and guard script: passed.
+  - `make pre-commit`: passed.
+
+- Issue #660 API-099 current slice:
+  - `cargo check -p rustfs --tests`: passed.
+  - `cargo fmt --all`: passed.
+  - `cargo fmt --all --check`: passed.
+  - `git diff --check`: passed.
+  - `bash -n scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_layer_dependencies.sh`: passed.
+  - Startup/init root compatibility consumer residual scan: passed.
+  - Root startup consumer wrapper residual scan: passed.
   - Rust risk scan on changed Rust files and guard script: passed.
   - `make pre-commit`: passed.
