+++ rustfs/src/startup_iam.rs
@@ -14,7 +14,7 @@
 
 use crate::app::context::AppContext;
 use crate::server::{ServiceStateManager, publish_ready_when_runtime_ready};
-use crate::storage_compat::ECStore;
+use crate::startup_storage_compat::ECStore;
 use rustfs_common::{GlobalReadiness, SystemStage};
 use rustfs_iam::init_iam_sys;
 use rustfs_kms::KmsServiceManager;
