+++ rustfs/src/startup_lifecycle.rs
@@ -17,7 +17,7 @@ use crate::{
     startup_iam::{IamBootstrapDisposition, publish_ready_for_iam_bootstrap},
     startup_services::StartupServiceRuntime,
     startup_shutdown::run_startup_shutdown_sequence,
-    storage_compat::ECStore,
+    startup_storage_compat::ECStore,
 };
 use rustfs_common::GlobalReadiness;
 use rustfs_scanner::init_data_scanner;
